+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [4.2.0] - 2025-11-12
+
+_Stable release based on [4.2.0-rc.1]._
+
 ## [4.2.0-rc.1] - 2025-11-10
 
 ### Added
@@ -270,6 +274,7 @@ _Stable release based on [0.1.0-rc.1]._
 
 - New changelog file.
 
+[4.2.0]: https://https://github.com/internetguru/laravel-feedback/compare/v4.1.0...v4.2.0
 [4.2.0-rc.1]: https://github.com/internetguru/laravel-feedback/releases/tag/v4.1.0
 [4.1.0]: https://https://github.com/internetguru/laravel-feedback/compare/v4.0.2...v4.1.0
 [4.1.0-rc.1]: https://github.com/internetguru/laravel-feedback/releases/tag/v4.0.2